divide tests for rust backend
